+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/CollectionCreationTimelineOrder.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class CollectionCreationTimelineOrder(val value: String) {
     OrderAdded("tweet_chron"), OldestFirst("curation_reverse_chron"), MostRecentFirst("tweet_reverse_chron")
