+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/EmbedAlign.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class EmbedAlign(val value: String) {
     Left("left"), Right("right"), Center("center"), None("none")
