+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/EmbedWidgetType.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class EmbedWidgetType(val value: String) {
     Video("video")
