+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/ListCreationMode.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class ListCreationMode(val value: String) {
     Public("public"), Private("private")
