+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/MediaCategory.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class MediaCategory(val value: String) {
     TweetImage("tweet_image"), TweetGif("tweet_gif"), TweetVideo("tweet_video"), AmplifyVideo("amplify_video")
