+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/SearchResultType.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class SearchResultType(val value: String) {
     Mixed("mixed"), Recent("recent"), Popular("popular")
