+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/TrendExclude.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class TrendExclude(val value: String) {
     Hashtags("hashtags")
