+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/UserStreamFilterLevel.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class UserStreamFilterLevel(val value: String) {
     None("none"), Low("low"), Medium("medium")
