+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/UserStreamReplies.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class UserStreamReplies(val value: String) {
     All("all")
