+++ src/main/kotlin/jp/nephy/penicillin/endpoints/parameters/UserStreamWith.kt
@@ -1,4 +1,4 @@
-package jp.nephy.penicillin.endpoints.parameter
+package jp.nephy.penicillin.endpoints.parameters
 
 enum class UserStreamWith(val value: String) {
     User("user"), Followings("followings")
